requirements.txt
paddleocr
fastapi
uvicorn
numpy
Pillow
